+++ module/Core/src/Action/AbstractTrackingAction.php
@@ -15,7 +15,6 @@ use Shlinkio\Shlink\Core\ShortUrl\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\ShortUrl\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\Core\ShortUrl\ShortUrlResolverInterface;
 use Shlinkio\Shlink\Core\Visit\RequestTrackerInterface;
-use Shlinkio\Shlink\IpGeolocation\Model\Location;
 
 abstract class AbstractTrackingAction implements MiddlewareInterface, RequestMethodInterface
 {
@@ -31,12 +30,9 @@ abstract class AbstractTrackingAction implements MiddlewareInterface, RequestMet
 
         try {
             $shortUrl = $this->urlResolver->resolveEnabledShortUrl($identifier);
-            $visit = $this->requestTracker->trackIfApplicable($shortUrl, $request);
+            $this->requestTracker->trackIfApplicable($shortUrl, $request);
 
-            return $this->createSuccessResp(
-                $shortUrl,
-                $request->withAttribute(Location::class, $visit?->getVisitLocation()),
-            );
+            return $this->createSuccessResp($shortUrl, $request);
         } catch (ShortUrlNotFoundException) {
             return $this->createErrorResp($request, $handler);
         }