+++ module/Core/test-db/Visit/Repository/VisitDeleterRepositoryTest.php
@@ -9,18 +9,26 @@ use Shlinkio\Shlink\Core\ShortUrl\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\ShortUrl\Model\ShortUrlCreation;
 use Shlinkio\Shlink\Core\ShortUrl\Model\Validation\ShortUrlInputFilter;
 use Shlinkio\Shlink\Core\ShortUrl\Resolver\PersistenceShortUrlRelationResolver;
+use Shlinkio\Shlink\Core\Visit\Entity\OrphanVisitsCount;
+use Shlinkio\Shlink\Core\Visit\Entity\ShortUrlVisitsCount;
 use Shlinkio\Shlink\Core\Visit\Entity\Visit;
 use Shlinkio\Shlink\Core\Visit\Model\Visitor;
+use Shlinkio\Shlink\Core\Visit\Repository\OrphanVisitsCountRepository;
+use Shlinkio\Shlink\Core\Visit\Repository\ShortUrlVisitsCountRepository;
 use Shlinkio\Shlink\Core\Visit\Repository\VisitDeleterRepository;
 use Shlinkio\Shlink\TestUtils\DbTest\DatabaseTestCase;
 
 class VisitDeleterRepositoryTest extends DatabaseTestCase
 {
     private VisitDeleterRepository $repo;
+    private ShortUrlVisitsCountRepository $visitsCountRepo;
+    private OrphanVisitsCountRepository $orphanVisitsCountRepo;
 
     protected function setUp(): void
     {
         $this->repo = $this->createRepository(Visit::class, VisitDeleterRepository::class);
+        $this->visitsCountRepo = $this->getEntityManager()->getRepository(ShortUrlVisitsCount::class);
+        $this->orphanVisitsCountRepo = $this->getEntityManager()->getRepository(OrphanVisitsCount::class);
     }
 
     #[Test]
@@ -51,12 +59,20 @@ class VisitDeleterRepositoryTest extends DatabaseTestCase
 
         $this->getEntityManager()->flush();
 
-        self::assertEquals(2, $this->repo->deleteShortUrlVisits($shortUrl1));
-        self::assertEquals(0, $this->repo->deleteShortUrlVisits($shortUrl1));
-        self::assertEquals(4, $this->repo->deleteShortUrlVisits($shortUrl2));
-        self::assertEquals(0, $this->repo->deleteShortUrlVisits($shortUrl2));
-        self::assertEquals(1, $this->repo->deleteShortUrlVisits($shortUrl3));
-        self::assertEquals(0, $this->repo->deleteShortUrlVisits($shortUrl3));
+        $this->assertVisitsDeletionForShortUrl($shortUrl1, 2);
+        $this->assertVisitsDeletionForShortUrl($shortUrl2, 4);
+        $this->assertVisitsDeletionForShortUrl($shortUrl3, 1);
+    }
+
+    private function assertVisitsDeletionForShortUrl(ShortUrl $shortUrl, int $expectedDeleteCount): void
+    {
+        // There should be at least one visit count before deletion
+        self::assertGreaterThan(0, $this->visitsCountRepo->count(['shortUrl' => $shortUrl]));
+        self::assertEquals($expectedDeleteCount, $this->repo->deleteShortUrlVisits($shortUrl));
+
+        // Visits counts are also deleted, and trying to delete again results in no visits deleted
+        self::assertEquals(0, $this->visitsCountRepo->count(['shortUrl' => $shortUrl]));
+        self::assertEquals(0, $this->repo->deleteShortUrlVisits($shortUrl));
     }
 
     #[Test]
@@ -72,7 +88,9 @@ class VisitDeleterRepositoryTest extends DatabaseTestCase
 
         $this->getEntityManager()->flush();
 
+        self::assertGreaterThan(0, $this->orphanVisitsCountRepo->count());
         self::assertEquals(6, $this->repo->deleteOrphanVisits());
         self::assertEquals(0, $this->repo->deleteOrphanVisits());
+        self::assertEquals(0, $this->orphanVisitsCountRepo->count());
     }
 }